+++ packages/zarm/src/pinch-zoom/interface.ts
@@ -1,18 +1,5 @@
-export interface Point {
-  clientX: number;
-  clientY: number;
-}
-
-export interface ApplyChangeOpts {
-  panX?: number;
-  panY?: number;
-  scaleDiff?: number;
-  originX?: number;
-  originY?: number;
-}
-
 export interface BasePinchZoomProps {
-  minScale: number;
-  maxScale: number;
-  onPinchZoom?: Function;
+  minScale?: number;
+  maxScale?: number;
+  onPinchZoom?: (scale: number, x: number, y: number) => void;
 }